condense SessionViewContainer.tsx to a single component - added dancer analytic mutation to update dancer analytics
